Add files via upload
Without interval means the pneumatic system will kept running once it is turned on. If the switch of channel 1/2 is on, the rheostat can be used to tune the output pressure. If the switch is off, the pneumatic system will generate a constant output that is set in the code.

With interval means the pneumatic system will rest for 5 seconds after running for 5 seconds. If the switch of channel 1/2 is on, the rheostat can be used to tune the output pressure. If the switch is off, the pneumatic system will generate a constant output that is set in the code.